WIP: use native LSP init message for handshake
This change updates the server to read its configuration options from
the initialization request's `initializationOptions.raMultiplex` field.
`initializationOptions` is usually reserved for LSP-specific
configuration, so this seemed like a good fit.

The workspace root path used is the same one as resolved by the client
and specified in the initialization request (although the field is
deprecated and I still need to handle using the new one if the old one
is absent).

I've tested this by configuring neovim to connect to
`ra-multiplex-server` directly:

    cmd = vim.lsp.rpc.connect('127.0.0.1', 27631),
    init_options = {
      raMultiplex = {
        server = "rust-analyzer"
      }
    },

It connects and works. Setting another `server` value here works as
intended too. This allows configuring a client to pass a different
server depending on filetype too (it should also be doable in a generic
way, instead of duplicating the configuration for each filetype/server).

There's still some missing work:

- The client portion of the changes need to be done (and these will be a
  bit more work). Essentially, I'd want the client to encapsulate its
  configuration into the initialization request. I mostly want some
  feedback on the approach before addressing this.
- I'm not sure what to do about the protocol and version fields in the
  init message, I think that it would make sense to drop them.
- `Init::from_reader` is now unused, so I think it should be dropped
  too.
- The `initializationOptions.raMultiplex` should be removed from the
  request before forwarding it to the server (just in case some server
  is picky about unknown fields).

Fixes: https://github.com/pr2502/ra-multiplex/issues/20