the big gateway shuffle
i'm sorry to all the open PRs about this, i'll port them myself.

the structure of the code was too confusing for too long, it was a relic
of having two binaries and trying to share as little as possible between
them to reduce code bloat. but now that we have a single binary it
doesn't make sense anymore and is very confusing.

this change will hopefully make the code more clear for anyone not
already familiar with it.